+++ ChangeLog.md
@@ -26,6 +26,7 @@ Incoming in 0.25.0
       `match().filter()` still returns a sequential array with reindexed values.
     * Removed `Stream.only()`. Use `Stream.limit().all()` instead.
     * Removed `IntStream.only()`. Use `IntStream.limit().all()` instead.
+    * Renamed `Group.textLength()` to `Group.length()`.
 * Bug fixes
     * Fixed a bug when using `Stream.values().keys().first()` didn't always reindex to `0`.
     * Fixed a bug when using regular groups in `stream()->asInt()` was allowed, but `usingDuplicateName()` groups